Removed left sidebar from non-front page on mobile.
If you're on a page that is NOT the front page, and you're on mobile
(i.e. the global-mobile-first CSS), then do NOT display the sidebar.
This seems to work pretty well! If you're on mobile, you get right to
the article, without the sidebar. If you're on a tablet, or a desktop
browser, you get 'the full layout.'