Removed non-mobile inverse settings.
It turns out that I do NOT need these CSS entries.

The comment references 'inverse' because the global (i.e.
mobile first) layout turns off or on certain sections, but these
settings turn them on or off (i.e. their opposite, or inverse,
setting). I thought I needed to do this for normal and wide, but
doing the inverse in the 'default' CSS is sufficient.